rename nexus example
