+++ typescript/graphql-nexus/schema.graphql
@@ -1,18 +1,22 @@
+### This file was generated by Nexus Schema
+### Do not make changes to this file directly
+
+
 """
 A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar.
 """
 scalar DateTime
 
 type Mutation {
-  createDraft(authorEmail: String!, data: PostCreateInput!): Post!
-  deletePost(id: Int!): Post!
-  incrementPostViewCount(id: Int!): Post!
+  createDraft(authorEmail: String!, data: PostCreateInput!): Post
+  deletePost(id: Int!): Post
+  incrementPostViewCount(id: Int!): Post
   signupUser(data: UserCreateInput!): User!
-  togglePublishPost(id: Int!): Post!
+  togglePublishPost(id: Int!): Post
 }
 
 type Post {
-  author: User!
+  author: User
   content: String
   createdAt: DateTime!
   id: Int!
@@ -33,9 +37,9 @@ input PostOrderByUpdatedAtInput {
 
 type Query {
   allUsers: [User!]!
-  draftsByUser(userUniqueInput: UserUniqueInput!): [Post!]
+  draftsByUser(userUniqueInput: UserUniqueInput!): [Post]
   feed(orderBy: PostOrderByUpdatedAtInput, searchString: String, skip: Int, take: Int): [Post!]!
-  postById(id: Int!): Post
+  postById(id: Int): Post
 }
 
 enum SortOrder {
