+++ typescript/graphql-nexus/src/server.ts
@@ -1,0 +1,15 @@
+import { ApolloServer } from 'apollo-server'
+import { schema } from './schema'
+import { context } from './context'
+
+const server = new ApolloServer({
+  schema: schema,
+  context: context,
+})
+
+server.listen().then(async ({ url }) => {
+  console.log(`\
+🚀 Server ready at: ${url}
+⭐️ See sample queries: http://pris.ly/e/ts/graphql#using-the-graphql-api
+  `)
+})
