+++ typescript/graphql/package.json
@@ -5,10 +5,10 @@
     "dev": "ts-node-dev --no-notify --respawn --transpile-only src/server",
     "start": "node dist/server",
     "clean": "rm -rf dist",
-    "build": "npm -s run clean && npm -s run generate && tsc",
-    "generate": "npm -s run generate:prisma && npm -s run generate:nexus",
-    "generate:prisma": "prisma generate",
-    "generate:nexus": "ts-node --transpile-only src/schema"
+    "build": "npm -s run clean && npm -s run generate && tsc && npm run build:schema",
+    "build:schema": "ts-node src/schema/index.ts",
+    "generate": "npm -s run generate:prisma",
+    "generate:prisma": "prisma generate"
   },
   "prettier": {
     "semi": false,
@@ -16,18 +16,19 @@
     "trailingComma": "all"
   },
   "dependencies": {
+    "@pothos/core": "^3.23.1",
+    "@pothos/plugin-prisma": "^3.37.0",
     "@prisma/client": "4.7.1",
-    "apollo-server": "3.11.1",
     "graphql": "16.5.0",
     "graphql-scalars": "1.20.1",
-    "nexus": "1.3.0"
+    "graphql-yoga": "^3.1.1"
   },
   "devDependencies": {
     "@types/node": "18.11.11",
     "prisma": "4.7.1",
     "ts-node": "10.9.1",
     "ts-node-dev": "2.0.0",
-    "typescript": "4.9.4"
+    "typescript": "4.9.3"
   },
   "prisma": {
     "seed": "ts-node --transpile-only prisma/seed.ts"
