+++ typescript/graphql/src/builder.ts
@@ -20,4 +20,7 @@ export const builder = new SchemaBuilder<{
   },
 })
 
+builder.queryType({})
+builder.mutationType({})
+
 builder.addScalarType('DateTime', DateTimeResolver, {})
