+++ typescript/graphql/src/schema/index.ts
@@ -1,0 +1,10 @@
+import { builder } from '../builder'
+import './user'
+import './post'
+import { writeFileSync } from 'fs'
+import { resolve } from 'path'
+import { printSchema } from 'graphql'
+
+export const schema = builder.toSchema({})
+
+writeFileSync(resolve(__dirname, '../../schema.graphql'), printSchema(schema))
